Refactored map marker wrappers.
