+++ lib/map/features/gym.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const Feature = require('./feature');
-const Settings = require('../settings');
-const FeatureLabel = require('./feature_label');
+const Settings = require('../../settings');
+const Marker = require('../../objects/marker');
+const MarkerLabel = require('../../objects/marker_label');
 
-class Gym extends Feature {
+class Gym extends Marker {
   constructor(options) {
     super(options);
     this.exEligibility = options.exEligibility.toLowerCase();
@@ -91,7 +91,7 @@ class Gym extends Feature {
         </div>
       </div>`;
 
-    return new FeatureLabel(this, this.marker, {
+    return new MarkerLabel(this, this.marker, {
       content: content,
       pixelOffset: new google.maps.Size(-7, -30)
     });
