+++ lib/map/features/pokestop.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const Feature = require('./feature');
-const Settings = require('../settings');
-const FeatureLabel = require('./feature_label');
+const Settings = require('../../settings');
+const Marker = require('../../objects/marker');
+const MarkerLabel = require('../../objects/marker_label');
 
-class Pokestop extends Feature {
+class Pokestop extends Marker {
   shouldShow() {
     return Settings.get('showPokestops');
   }
@@ -22,7 +22,7 @@ class Pokestop extends Feature {
         </div>
       </div>`;
 
-    return new FeatureLabel(this, this.marker, {
+    return new MarkerLabel(this, this.marker, {
       content: content,
       pixelOffset: new google.maps.Size(0, -30)
     });
