+++ lib/map/features/portal.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const Feature = require('./feature');
-const Settings = require('../settings');
-const FeatureLabel = require('./feature_label');
+const Settings = require('../../settings');
+const Marker = require('../../objects/marker');
+const MarkerLabel = require('../../objects/marker_label');
 
-class Portal extends Feature {
+class Portal extends Marker {
   shouldShow() {
     return Settings.get('showPortals');
   }
@@ -22,7 +22,7 @@ class Portal extends Feature {
         </div>
       </div>`;
 
-    return new FeatureLabel(this, this.marker, {
+    return new MarkerLabel(this, this.marker, {
       content: content,
       pixelOffset: new google.maps.Size(0, -30)
     });
