+++ lib/map/features/s2_cell.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const Polygon = require('./polygon');
-const Settings = require('./../settings');
+const Settings = require('../../settings');
+const Polygon = require('../../objects/polygon');
 
 class S2Cell extends Polygon {
   constructor(s2CellId) {
