+++ lib/map/map.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const Settings = require('./settings');
-const Utils = require('./utils');
-const SearchBox = require('./search_box');
-const Gym = require('./features/gym');
-const Pokestop = require('./features/pokestop');
+const Settings = require('../settings');
+const Utils = require('../utils');
+const SearchBox = require('../search_box');
 const Portal = require('./features/portal');
+const Pokestop = require('./features/pokestop');
+const Gym = require('./features/gym');
 const S2Cell = require('./features/s2_cell');
 
 window.initMap = () => {
