+++ lib/nests/nest.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const Utils = require('../utils');
-const Polygon = require('./polygon');
-const FeatureLabel = require('./feature_label');
+const Polygon = require('../objects/polygon');
+const MarkerLabel = require('../objects/marker_label');
 
 class Nest extends Polygon {
   constructor(options) {
@@ -96,7 +96,7 @@ class Nest extends Polygon {
       clickTargets.push(this.marker);
     }
 
-    return new FeatureLabel(this, clickTargets, {
+    return new MarkerLabel(this, clickTargets, {
       content: content
     });
   }
