+++ lib/nests/nests.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const Settings = require('./settings');
-const Utils = require('./utils');
-const SearchBox = require('./search_box');
-const Nest = require('./features/nest');
+const Settings = require('../settings');
+const Utils = require('../utils');
+const SearchBox = require('../search_box');
+const Nest = require('./nest');
 
 window.initMap = () => {
   const latLngStr = Utils.urlParameter('ll');
