+++ lib/objects/marker.js
@@ -2,9 +2,9 @@
 
 const Utils = require('../utils');
 const Settings = require('../settings');
-const FeatureLabel = require('./feature_label');
+const MarkerLabel = require('./marker_label');
 
-class Feature {
+class Marker {
   constructor(options) {
     this.options = options;
     this.id = options.id;
@@ -78,8 +78,8 @@ class Feature {
 
   _createLabel() {
     const content = `
-      <div class="feature-label">
-        <div class="feature-label-name">
+      <div class="marker-label">
+        <div class="marker-label-name">
           <b>${this.name}</b>
         </div>
       </div>`;
@@ -91,4 +91,4 @@ class Feature {
   }
 }
 
-module.exports = Feature;
+module.exports = Marker;
