+++ lib/objects/marker_label.js
@@ -1,8 +1,8 @@
 'use strict';
 
-class FeatureLabel {
-  constructor(feature, clickTarget, options) {
-    this.feature = feature;
+class MarkerLabel {
+  constructor(marker, clickTarget, options) {
+    this.marker = marker;
 
     if (Array.isArray(clickTarget)) {
       this.clickTargets = clickTarget;
@@ -25,7 +25,7 @@ class FeatureLabel {
   }
 
   open(map) {
-    this.label.setPosition(this.feature.location);
+    this.label.setPosition(this.marker.location);
     this.label.open(map);
     this.isOpen = true;
   }
@@ -44,4 +44,4 @@ class FeatureLabel {
   }
 }
 
-module.exports = FeatureLabel;
+module.exports = MarkerLabel;
