+++ lib/objects/polygon.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Utils = require('../utils');
-const FeatureLabel = require('./feature_label');
+const MarkerLabel = require('./marker_label');
 
 class Polygon {
   constructor(options) {
@@ -98,7 +98,7 @@ class Polygon {
         </div>
       </div>`;
 
-    return new FeatureLabel(this, this.polygon, {
+    return new MarkerLabel(this, this.polygon, {
       content: content
     });
   }
