+++ webpack.config.js
@@ -2,8 +2,8 @@ module.exports = {
   entry: {
     sign_in: './lib/sign_in.js',
     sign_in_token: './lib/sign_in_token.js',
-    map: './lib/map.js',
-    nests: './lib/nests.js'
+    map: './lib/map/map.js',
+    nests: './lib/nests/nests.js'
   },
   output: {
     path: `${__dirname}/assets/js`,
